+++ src/Resources/Policy.Autorest/custom/Remove-AzPolicyExemption.ps1
@@ -49,7 +49,7 @@ param(
     [Microsoft.Azure.PowerShell.Cmdlets.Policy.Category('Path')]
     [System.String]
     # The scope of the policy exemption.
-    # Valid scopes are: management group (format: '/providers/Microsoft.Management/managementGroups/{managementGroup}'), subscription (format: '/subscriptions/{subscriptionId}'), resource group (format: '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}', or resource (format: '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/[{parentResourcePath}/]{resourceType}/{resourceName}'
+    # Valid scopes are: management group (format: '/providers/Microsoft.Management/managementGroups/{managementGroup}'), subscription (format: '/subscriptions/{subscriptionId}'), resource group (format: '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}'), or resource (format: '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/[{parentResourcePath}/]{resourceType}/{resourceName}')
     ${Scope},
 
     [Parameter(ParameterSetName='Id', Mandatory, ValueFromPipelineByPropertyName)]
@@ -73,12 +73,6 @@ param(
     [System.Management.Automation.SwitchParameter]
     # When $true, skip confirmation prompts
     ${Force},
-
-    [Parameter()]
-    [Obsolete('This parameter is a temporary bridge to new types and formats and will be removed in a future release.')]
-    [System.Management.Automation.SwitchParameter]
-    # Causes cmdlet to return artifacts using legacy format placing policy-specific properties in a property bag object.
-    ${BackwardCompatible} = $false,
 
     [Parameter()]
     [Alias('AzureRMContext', 'AzureCredential')]
@@ -145,7 +139,6 @@ begin {
 
     $mapping = @{
         Delete = 'Az.Policy.private\Remove-AzPolicyExemption_Delete';
-        DeleteViaIdentity = 'Az.Policy.private\Remove-AzPolicyExemption_DeleteViaIdentity';
     }
 }
 
@@ -180,18 +173,6 @@ process {
         $null = $PSBoundParameters.Remove('Force')
     }
 
-    # use passthru for backward compatibility with previous SDK cmdlets: remove cmdlet always returned a value
-    if ($PSBoundParameters.ContainsKey('PassThru')) {
-        $BackwardCompatible = $PassThru
-        $PSBoundParameters.PassThru = $PassThru
-    }
-    elseif ($BackwardCompatible) {
-        $PSBoundParameters.PassThru = $BackwardCompatible
-    }
-
-    # remove non-generated parameters
-    $null = $PSBoundParameters.Remove('BackwardCompatible')
-
     # remove the exemption if inputs resolve and user confirms
     if ($resolved.Scope -and $PSCmdlet.ShouldProcess($target)) {
 
@@ -212,15 +193,8 @@ process {
         $result = Invoke-Command -ScriptBlock $scriptCmd
     }
 
-    # return result of remove
-    if ($BackwardCompatible) {
-        if ($result -is [boolean]) {
-            $PSCmdlet.WriteObject($result)
-        }
-        else {
-            # $result is not a boolean for some reason, so return constant $true
-            $PSCmdlet.WriteObject($true)
-        }
+    if ($PassThru) {
+        $PSCmdlet.WriteObject($result)
     }
 }